Added PerformerByURL scraper, feminized.com to FemdomEmpire
feminized.com is run by the same people as femdomempire, using the same
CMS.

From their logo it's clear the site's name is Femdom Empire, not FemDom
Empire, so I renamed the yml file too.

I thought about adding performer stats to the scraper, but it's pretty
hard due to the way their tags work. Possible with XPath 2.0 (and heavy
use of regexes) but only present on Femdom Empire - and I'm not really
bothered about it, so I haven't looked into whether the XPath parser in
the scraper supports XPath 2.