replace scraper for SARJ LLC with a python script (#414)
Replace the JSON based scraper for the SARJ LLC websites with a more
elabore script.

The scripts supports the same features. Including remapping site ids to
their respective names, supporting the same sites, mapping the same
fields etc.. But also supports some additional features, which are
scraping performers by URL, and scraping scenes and galleries by their
fragments based on name and date. This thus allows to fill in the title
and date for a scene, applying the scraper, and all other information
will automatically be scraped. When scraping information the URL is also
automatically updated. So for example when scraping a scene released by
Viv Thomas but using the metartnetwork.com URL the URL will
automatically be replaced by the URL on vivthomas.com.

Using this script based scraper would also allow to add additional
features. Like a more elaborate search by using the search API and using
additional fields from the fragment. But this is more complex as for one
the fragment contains publishes ids (like performer "1", not including
the name), so that information should then first be fetched using the
GraphQL API. And the search API of the site seems to be rather picky in
the search terms.