sarj-lc: fix scraping models by URL
In some cases scraping a model by just using the
/api/model?namemodel-name endpoint doesn't work and returns a model not
found error, until opening the model page in the browser. After
investigation it seems like it is both required to transform the model
name by capitalizing the parts and replacing the dashes with a space
(code should be the Python equivalent of how the site does it in
JavaScript), and to add the order and direction arguments. With these
changes loading the model doesn't result in a 404 anymore.