sarj-llc: support searching models by name
Add support for searching models by name. This is kind of a hack as you
can't really search models. You can however run a search, of which the
result will include items of the model types. So this iterates over all
search results (including all pages) and filters out the models. When a
search phrase has lots of results it will inevitable take a couple of
seconds to iterate all pages.

Navigation happens using 30 results per page, equaling the website.
Obviosly this could be raised. And as model results always seem to be
last it might even be worthwile to start iterating from the last result
(i.e.: make a call with pageSize=0 to get the total count, and fetch the
last page first and then decrement the page).