Update IFeelMyself.py: search results beyond 2022
The cookie "ifm_prefs" in functions scrapeScene and queryPerformer contains a hardcoded end date of October 2022, which results in search results after October 2022 to not appear in queries.

The three instances of the "ifm_prefs" cookie all have the same value set, a URL-encoded dictionary which sets search parameters. I decoded the string, changed the "to_month" value from 10 to 12 and the "to_year" value from 2022 to 3000. If a greater value beyond the year 3000 is desired I would be happy to modify my PR.

In testing, both in-browser and in the script, using a "to_year" value of 3000 did not result in any dicernable breakages. Results from January 2006 populated, as well as results through January 31st 2023 (in separate searches).