Update and rename DoTheWife.yml to SmutPuppet.yml
Repurposing the "Do the Wife" scraper for additional sites in the SmutPuppet network.
- URLs have been added for the other SmuttPuppet sites.
- Studio name is being grabbed from alt text for studio image.
- Image URLs are being built from smutpuppet domain

The above was tested with a random scene URL from each site. 

Performer by URL scraper added.  Not much performer info is available, just the image, date of birth and nationality.  Nationality is a bit of a mess.  Most are the country name, but not all.  The mapping is for the ones that I was able to find as I browsed through performers before my eyes glazed over too much.  Some are in ALL CAPS, others in a more sensible display.