New scraper for Jacquie et Michel Elite
Added new URL scraper for scenes and movies for Jacquie et Michel Elite (https://www.jacquieetmichelelite.com/en/).

After seeing that some class values had trailing spaces within the double quotes, it was decided to be super paranoid and do all @class matching using "contains".

The Movie title on the Scenes scraper can only pull the value in lower case, since the movie itself is only available once as part of a URL.  This can be corrected easily by scraping the movie itself once it's created.

No tags available on the studio site, unfortunately.

The way that the site lays out their movies is the same as how they lay out their scenes, so the scene scraper can also be used on the URL for a movie. This is handy if you happen to have the movie as one file.  It's for this reason that the scene scraper URL is more general than one would initially believe is necessary.