Update KBProductions.py to work with lucidflix.com
Added 'episodes' to URL regex as Lucidflix uses /episodes/ in their scene URLs.
The scraper pick up "Lucid flix" as the studio name, as it's scraping from the only portion of the page where it's referenced as such, where it's LucidFlix absolutely everywhere else.  I've added a studioMap in the scrape_scene function to fix this, and to allow for easy updates for any future sites where they do something equally silly.