Update Andomark for new URLs
Added URLs that are compatible with this scraper.
Added Studio map, with an attempt to match the mapped name to how the studio name is presented in the site HTML
Modified image processing to maximize the chance of getting a suitable image.
Removed one site that no longer appears to exist.