Update TugPass.yml
Added a number of sites that are in the TugPass network, or which otherwise use the TugPass template.

Reworked image selector, as there are a number of different ways that sites reference images.  This is for both existing and new sites in the scraper.

Added studio map for those that use a slightly different presentation than the rest in the Studio selector.