Update PegasProductions.yml
Added User-Agent header to prevent 403-Forbidden response.

Added scraper for second format that studio has for scenes.  It's less useful than their primary format, but adding this in the event that a scene is only available via the newer, less useful, format.

Changed studio name to match what StashDB uses for the studio.

Added Director field to the scraper for the primary format.

Added French support to the scraper for the primary format should someone prefer to scrape this site in French rather than English.