Add studio map for Hitzefrei.yml (#1742)
* Update Hitzefrei.yml

The studio has several sub-studios / series for their content, each with a subdomain URL from hitzefrei.com, and one of which at this time is present in StashDB as a sub-studio of Hitzefrei.  This change adds a studio map to the scraper which will provide the series name as the studio when scraping from the series' URLs.  The content is also available from the hitzefrei.com URL and use the 'tour' subdomain, but unfortunately there is at present no indication of the series when viewed this way.  Thus scenes scraped directly from hitzefrei.com will be given the Hitzefrei studio.

Studio map:
          citycheck: City Check
          cuffemall: Cuff 'em All
          familyaffairs: Family Affairs
          fanalarm: Fanalarm
          fuckonarrival: Fuck On Arrival
          milfhunters: Milf Hunters
          pattisanals: Milf Hunters
          tour: Hitzefrei
          unleashed: Unleashed

* Update Hitzefrei.yml

Corrected mappings for pattisanals and milfhunters.