Tug pass updates (#1714)
* Update TugPass.yml

Added a number of sites that are in the TugPass network, or which otherwise use the TugPass template.

Reworked image selector, as there are a number of different ways that sites reference images.  This is for both existing and new sites in the scraper.

Added studio map for those that use a slightly different presentation than the rest in the Studio selector.

* Delete scrapers/Mylked.yml

This is a site in the TugPass network.  Rolling this scraper into TugPass.

* Update SCRAPERS-LIST.md

Added several new sites that use the TugPass scraper.

Updated mylked and teasepov to use TugPass scraper.

* Delete scrapers/TeasePOV.yml

This site is part of the TugPass network.  Rolling scraper into TugPass.

---------

Co-authored-by: Maista <131594657+Maista6969@users.noreply.github.com>