Create Pornmuschimovie.yml
New URL YML scene scraper for sites in the Pornmuschimovie network:

      - dirty-chubby.com
      - dirtycasting.com
      - lustyxxxfamily.com
      - pornmuschimovie.com
      - sexgamesprivate.com
      - xxxrealityfuck.com

Notes:
No scenes appear to have performer information
The sites are a bit sloppy, with some scenes having their descriptions entered as tags.  If you notice very strange tags it's probably not because the selectors are not the proper ones to use, but it's more likely that the site messed up in building those scenes.