Update XConfessions.yml
Modified image selector to attempt several methods to find a suitable picture.  
First, look in the script for a cover_title_picture value.  

If that's not present, look for two consecutive images with UIDs in the URL.  The first is the movie poster, the second is the scene image.

If that's not present, look for a time stamp followed closely by two images.  Again the first is the movie version, the second is the image.

If we still don't have a match, look for a confession_cover_image value.  These are pretty good, but won't have the scene name.

Finally, if nothing matches, try the last image in the script.  It's often the right one.

Yeah...I know...weird.  But this reliably snags a suitable image for the scene, though often it will be the vertical movie poster instead of a horizontal scene poster.  A solution to improve this may yet present itself in the future.

Other changes: 

Modified movie Front Image selector to grab the vertical image more suitable for movie poster images.

Passing along the URL when creating movies when scenes are scraped.

Replaced " for ' for strings, for consistency.