Update Vivid.yml
Removed vivid.com, as it is now scrapable by the Adultime network scraper.

While all the remaining URLs in this scraper now redirect to vivid.com, links to individual scenes (as found on StashDB for some scenes) are still serving data and are scrapable by this scraper.  As such this will remain behind with those URLs until scraping breaks.