Update SCRAPERS-LIST.md
Added many new sites that use the ModelCentroAPI scraper, and remove a few that no longer exist or have no scrapable content.

Too many sites to manually edit.  The entire list was passed through a sort, so there may be some reordering of sites not otherwise related to this change.