ATKGirlfriends Try Catch Addition for Missing Elements (#1883)
* Add a Try Catch

Added a Try Catch to the code. Sometimes these elements are not present causing errors to be thrown or scenes not be scrapped. Appears to work in the two test cases I preformed

* Height Fix

Even the Height Is blank

* Fix for Compilation Scenes

The scraper now handles Compilation Scenes as well. It returns no model unfortunately but determine the model is out of my depth. There may be a way through the Title description but I couldn't think of a way for that to be accurate