Update GirlsOutWest.yml for latest site layout
Girls Out West has changed their layout, meaning a rewrite of the scraper is needed.

Scene and Performer URL scrapers have been modified to scrape the new layout.

It appears that their galleries are now behind a paywall, so it is not possible to scrape those.  I'd be happy to be proven wrong and return gallery scraping if a public URL can be provided.