Update GirlsOutWest.yml for latest site layout (#1889)
* Update GirlsOutWest.yml for latest site layout

Girls Out West has changed their layout, meaning a rewrite of the scraper is needed.

Scene and Performer URL scrapers have been modified to scrape the new layout.

It appears that their galleries are now behind a paywall, so it is not possible to scrape those.  I'd be happy to be proven wrong and return gallery scraping if a public URL can be provided.

* Update GirlsOutWest.yml

Added gallery scraping.
Added missing details to scene scraping.
Fixed typos in "last updated" line.

* Update GirlsOutWest.yml

Fixed silly typo.  Why, oh why, can't 'join' be an alias for 'concat'?

* Update GirlsOutWest.yml

When not in 'vi', don't try to use 'vi' commands.