Fix Putalocura Details (#1928)
* Fix Putalocura Details

The details in the scenes on Putalocura usually include links to the performer and other scenes, so the current scraper using the text() function will stop parsing the contents of the detail <p> tag when a link tag <a> is hit. Fix is to remove the text() specifier to allow the complete <p> tag to be automatically parsed. Issue can be seen with this link https://www.putalocura.com/en/spanish-glory-hole/his-first-glory-hole-la-sadica-vive and subsequent fix

* Update Date on Putalocura Details fix

Update the