Rule34Video: Fix video description sometimes not getting read properly (#1965)
* Rule34Video: Fix video description sometimes not getting read properly

Fixed an issue where the video description would miss line-breaks and would read wrong elements when empty. 
Would've wanted to insert correct linebreak ammount for mirroring the correct formating off of the site (e.g. multiple linebreaks directly after one-and-other), but my skills are too limited for that

* Small typo correction

Corrected a small typo in the date