Allow PremiumBukkake to scrape all scenes type from any of the site URL's
There are 3 PremiumBukkake sites, a free site, a paywalled site(not protected)  and a members site(protected). The free site is published after the other sites and not all scenes have are published, the First & Second camera scenes and some of the BTS and Interviews never make it to this free site. The free site title is not always consistent with the main sites and the release date lags as it is delayed by a few days from the main sites. Prior to this if it wasn't on the free site it could not be scraped. This modification allows you to supply any of the 3 site scene URLs (calculating the other 2) and allows any missing scenes (on the free site) to be scraped, also correcting inconsistencies in the title ands the dates using the info from the main paywalled site.