Update Algolia.py - MAIN_CHANNELS_AS_STUDIO_FOR_SCENE logic
Changed logic for MAIN_CHANNELS_AS_STUDIO_FOR_SCENE processing in the determine_studio_name_from_json subroutine, as the check if a site is in the dictionary for studio name processing was not using the proper key.  The value returned in the JSON for this key is a dictionary itself.  Changed the check to match the SITES_SEGMENT_USING_MAIN_CHANNEL_AS_SCENE_STUDIO check in the same block.

Added some debugging lines so it's more obvious which IF block is being triggered to assist in future efforts to hunt down studio name processing issues.

Added BiPhoria to MAIN_CHANNELS_AS_STUDIO_FOR_SCENE  dictionary to allow the scraper to return BiPhoria as the studio instead of  the scene name.