Improvements, Make URL compatible with JF 10.9+
Jellyfin 10.9 changed the Way Items can be referenced by URL. While the old pattern still works, it is deprecated now and will be removed in the future. Therefore the URL-Syntax has been updated to match the new Standard.
Furthermore I implemented two configurable Options in the config.ini:
If "Scene2Movie" is set to true, an Movie will be automatically created when scraping a new Scene. This is helpful if you have mainly Movies in Jellyfin and wants to create Movies in Stash accordingly.
If "IgnoreSceneImage" ist set to true, the Primary Image for the Scene will not be pulled from Jellyfin. This is helpful if you have Coverart in Jellyfin for your Movies, but want to make sure that you only use it for movies, not for the Scenes