Update vixenNetwork.py
Added Wifey to the studios dictionary.

Changed "isValidURL" processing to use split instead of lstrip and rstrip, as the previous implementation coded under a wrong assumption on how the strip functions work, which resulted in the 'w' from wifey from being stripped.  The new implementation splits on '.' and takes the second element, so www.wifey.com becomes "wifey" instead of "ifey".