Update GroobyNetwork-Partial.yml to handle performer names in titles (#2336)
For several sites, when there are multiple performers, the scene title will include the performer names and then the title. This causes the xpath to just result in a title of `&`.

Examples:
* https://www.transnificent.com/tour/trailers/Felix-Raunchy-Job-Interview.html
* https://www.realtgirls.com/tour/trailers/Probation-Officer-Jenna-Jaden-Gets-Attitude-Adjustment-From-Smash-Thompson.html

To fix this, I've adjusted the xpath query to only take the *last* text element when matching the `trailer_toptitle_left` element. I manually tested this against all the other sites for this scraper, to confirm that this change doesn't break their ability to scrape the scene title.