Add scraper for PaysiteManager sites (#2333)
* Create PaysiteManager.yml

Add URL Scene Scraper for sites in the Paysite Manager network.  Based on the existing Shinybound and FetishPro scrapers, with updated selectors to cover the differences in some of the sites, particularly in tags and images, and to add the Studio Code based on the "cart item".

* Delete scrapers/ShinyBound.yml

URLs covered by new PaysiteManager scraper.

* Delete scrapers/FetishPro.yml

Site covered by new PaysiteManager scraper.