Add Frenchtwinks studio code and fix details
This fixes an issue that existed when scraping the details where linebreaks were lost and the subtitle appeared smashed at the start of the description. This changes it to properly pull line breaks so the details match what is displayed visually on the site.

This also adds studio codes for French Twinks. These can be verified by looking at the `contentUrl` in the JSON-LD for the `VideoObject`.