Correct scene cover image
The previous image that was being used was the open graph image, which in many cases matches a full size scene cover image but in a handful of cases gives a very different one. This is because the open graph image that was being used is the first gallery image from the video.

This changes it to build the scene cover based on the chosen gallery image. This is available as the `thumbnailUrl` in the JSON-LD block for the `VideoObject`, but the link there is a compressed thumbnail of the one used within the scene player. This uses the full size scene image instead.

Note that French Twinks randomizes the actual screen cover displayed. As a result, it makes sense to standardize on the scene cover matching the JSON-LD thumbnail as that is at least static.