PremiumBukkake.com ensure valid performers
Issue
The PB scraper uses 2 PB sites to build information from, if one of the sites does not exist it falls back on information from the second site. In cases where a performer does not exist on the initial site and there is no URL for the scene on the second site, the performer field had a single entry of "None" which caused "The requested element is null which the schema does not all" error in stash interface.

Fix
Before assigning the scraped performer array add a check to see if it is valid.