Added SearchByURL to JAVDatabase and replaced reference to 404 scraper
sceneScraper references were returning error 404 not found when a url wasn't available. sceneSearch constructs the url based on the JAV code and as such implicit url defeats the purpose of the scraper. Added QueryFragment utility for when the filename parser fails to find a valid JAV code automatically, so user may search for video data by using a manually-entered JAV code.